Prg7.js
Create a simple REST API using Express.js and PostgreSQL that allows users to interact 
with a database. The API should support the following actions: 
a. Create: Add a new user to the database. 
b. Read: Retrieve all users from the database. 
c. Update: Update a user's information in the database. 
d. Delete: Delete a user from the database.